requirements_dev.txt
bumpversion==0.5.3
wheel==0.23.0
watchdog==0.8.3
flake8==2.4.1
tox==2.1.1
coverage==4.0
Sphinx==1.3.1
cryptography==1.0.1
PyYAML==3.11

apache-libcloud
flask_restful
flask-restful-swagger